# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise Asset Management API
  version: 1.0.0
  description: >
    Enterprise-grade multi-tenant asset lifecycle management API.
    Designed for backend architecture and automation practice.
servers:
  - url: http://localhost:8080
    description: Local Development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ExecutiveDashboard:
      type: object
      properties:
        totalAssets:
          type: integer
        activeAssets:
          type: integer
        maintenanceAssets:
          type: integer
        retiredAssets:
          type: integer
paths:
...
  /api/dashboard/executive:
    get:
      tags: [Dashboard]
      summary: Dashboard executivo (ADMIN)
      responses:
        200:
          description: Dados consolidados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutiveDashboard'
        403:
          description: Acesso negado